added default platform values for known properties
currently when parsing a github or bitbuck or any other known platform
the _user property is not being set. this is because it is not in the
pattern/format since it is always known. i added a way for setting
default property values for known platform properties. i also updated
the test and ran them. all passed.